requirements.txt
pyspark==3.5.0
pandas>=2.0.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
numpy>=1.24.0
openai>=1.0.0
PyPDF2>=3.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
pydantic>=2.0.0
